Fixed spec. Refactored partials for edit visit form.
