+++ app/helpers/study_schedule_helper.rb
@@ -54,16 +54,4 @@ module StudyScheduleHelper
 
     options_for_select(options)
   end
-
-  def move_to_position_options_for(selected)
-    selected.arm.visit_groups.where.not(id: selected.id).map do |visit_group|
-      ["before " + visit_name_with_day(visit_group), visit_group.position < selected.position ? visit_group.position : visit_group.position - 1]
-    end << ["as last", selected.arm.visit_groups.size]
-  end
-
-  private
-
-  def visit_name_with_day(visit_group)
-    visit_group.name + (!visit_group.day.nil? ? " / Day #{visit_group.day}" : "")
-  end
 end
