+++ app/views/study_schedule/management/manage_visits/_add_visit_form.html.haml
@@ -30,7 +30,7 @@
             .form-group
               = a.label "position", t(:visit_groups)[:position], class: "col-sm-4 control-label"
               .col-sm-7
-                = a.select "position", options_from_collection_for_select(arm.visit_groups, 'position', 'insertion_name') << vg_add_as_last_option(arm), {include_blank: ''}, class: "form-control selectpicker"
+                = a.select "position", options_for_select(insert_to_position_options(arm.visit_groups)), { include_blank: '' }, class: "form-control selectpicker"
             .form-group
               = a.label "window_before", t(:visit_groups)[:window_before], class: "col-sm-4 control-label"
               .col-sm-7
