+++ app/views/study_schedule/management/manage_visits/_navigate_visit_form.html.haml
@@ -20,7 +20,7 @@
           .form-group
             = label_tag 'vg_form_label', t(:visit)[:object], class: "col-sm-4 control-label"
             .col-sm-7
-              = select_tag "vg_form_select", options_from_collection_for_select(arm.visit_groups, 'id', 'name', visit_group.id), class: "selectpicker form-control"
+              = select_tag "vg_form_select", options_for_select(edit_visit_options(arm.visit_groups), visit_group.id), class: "selectpicker form-control"
       - if intended_action == "edit"
         %hr
           .edit_visit_area
