+++ spec/features/study_schedule/management/identity_edits_visit_groups_spec.rb
@@ -7,7 +7,7 @@ feature 'Identity edits visit groups for a particular protocol', js: true do
       scenario "and sees the visit group on the arm" do
         given_i_am_viewing_an_arm_with_multiple_visit_groups
         when_i_click_the_add_visit_group_button
-        when_i_fill_in_the_form
+        when_i_fill_in_the_form(day: 10000)
         when_i_click_the_add_submit_button
         then_i_should_see_the_visit_group
       end
@@ -15,7 +15,7 @@ feature 'Identity edits visit groups for a particular protocol', js: true do
       scenario "and sees a flash notification" do
         given_i_am_viewing_an_arm_with_multiple_visit_groups
         when_i_click_the_add_visit_group_button
-        when_i_fill_in_the_form
+        when_i_fill_in_the_form(day: 1000)
         when_i_click_the_add_submit_button
         then_i_should_see_a_flash_message_of_type 'add'
       end
@@ -27,8 +27,7 @@ feature 'Identity edits visit groups for a particular protocol', js: true do
         @original_visit_group_2 = @arm.visit_groups.second
 
         when_i_click_the_add_visit_group_button
-        when_i_fill_in_the_form
-        when_i_set_the_position_to "insert before #{@arm.visit_groups.second.name}"
+        when_i_fill_in_the_form(position: "before #{@arm.visit_groups.second.name}")
         when_i_click_the_add_submit_button
         then_i_should_see_the_position_is 1
       end
@@ -41,7 +40,7 @@ feature 'Identity edits visit groups for a particular protocol', js: true do
         given_i_am_viewing_an_arm_with_multiple_visit_groups
         when_i_click_the_edit_visit_group_button
         when_i_set_the_name_to 'VG 2'
-        when_i_set_the_day_to 44
+        when_i_set_the_day_to 2
         when_i_click_the_save_submit_button
         then_i_should_see_the_updated_visit_group
       end
@@ -149,11 +148,11 @@ feature 'Identity edits visit groups for a particular protocol', js: true do
     wait_for_ajax
   end
 
-  def when_i_fill_in_the_form
+  def when_i_fill_in_the_form(opts = {})
     bootstrap_select "#visit_group_arm_id", "#{@arm.name}"
-    fill_in "visit_group_name", with: "VG"
-    fill_in "visit_group_day", with: "4"
-    bootstrap_select "#visit_group_position", "Add as last"
+    fill_in "visit_group_name", with: opts[:name] || "VG"
+    fill_in "visit_group_day", with: opts[:day] || "13"
+    bootstrap_select "#visit_group_position", opts[:position] || "as last"
     wait_for_ajax
   end
